% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_noMed.R
\name{generate_data_noMed}
\alias{generate_data_noMed}
\title{generate_data_noMed}
\usage{
generate_data_noMed(
  J = 100,
  njrange = c(50, 100),
  Yfamily = "binomial",
  if.null = FALSE,
  seed = 123456,
  num_x = 3,
  a_on_x = sqrt(0.05625/3),
  a_on_z = sqrt(0.15/1),
  x_z = 0,
  yintercept = 1,
  y_on_a = 0.5,
  y_on_az = 0.2,
  y_on_anj = 0.2,
  y_on_x = sqrt(0.15/num_x),
  y_on_z = sqrt(0.4),
  quadratic.A = FALSE,
  quadratic.Y = FALSE,
  iccx = 0.2,
  icca = 0.2,
  iccy = 0.2,
  include_truevals = TRUE,
  include_overlapMsg = TRUE,
  plot_PSdiagnostics = FALSE,
  randomize = FALSE,
  ensure_cluster_positivity = TRUE
)
}
\arguments{
\item{J}{Integer. Number of clusters (default: 100).}

\item{njrange}{Integer vector of length 2. Range (min, max) for cluster sizes (default: \code{c(50, 100)}).}

\item{Yfamily}{Character. Family for the outcome (\code{"gaussian"} or \code{"binomial"}). Defaults to \code{"binomial"}.}

\item{if.null}{Logical. If \code{TRUE}, generate data under the null hypothesis (i.e., no effects). Defaults to \code{FALSE}.}

\item{seed}{Integer. Random seed for reproducibility (default: 123456).}

\item{num_x}{Integer. Number of individual-level covariates (\code{X}). Defaults to 3.}

\item{a_on_x}{Numeric. Effect of \code{X} on treatment \code{A}. Defaults to \code{sqrt(0.05625 / 3)}.}

\item{a_on_z}{Numeric. Effect of \code{Z} on treatment \code{A}. Defaults to \code{sqrt(0.15 / 1)}.}

\item{x_z}{Numeric. Correlation between \code{X} and the cluster-level variable \code{Z} (default: 0).}

\item{yintercept}{Numeric. Intercept in the outcome \code{Y} model (default: 1).}

\item{y_on_a}{Numeric. Main effect of \code{A} on \code{Y} (default: 0.5).}

\item{y_on_az}{Numeric. Interaction effect of \code{A} and \code{Z} on \code{Y} (default: 0.2).}

\item{y_on_anj}{Numeric. Interaction effect of \code{A} and cluster size (\code{nj}) on \code{Y} (default: 0.2).}

\item{y_on_x}{Numeric. Effect of each individual-level covariate \code{X} on \code{Y}.
Defaults to \code{sqrt(0.15 / num_x)}.}

\item{y_on_z}{Numeric. Effect of \code{Z} on \code{Y}. Defaults to \code{sqrt(0.4)}.}

\item{quadratic.A}{Logical. If \code{TRUE}, include quadratic term for \code{A} in the mediator/outcome models (default: FALSE).}

\item{quadratic.Y}{Logical. If \code{TRUE}, include quadratic term for \code{Y} itself (used in some advanced simulations) (default: FALSE).}

\item{iccx}{Numeric. Intra-class correlation for the \code{X} variables (default: 0.2).}

\item{icca}{Numeric. Intra-class correlation for treatment \code{A} (default: 0.2).}

\item{iccy}{Numeric. Intra-class correlation for outcome \code{Y} (default: 0.2).}

\item{include_truevals}{Logical. If \code{TRUE}, call \code{trueVals2.0d()} to compute the true potential outcomes and
mediation effects (default: TRUE).}

\item{include_overlapMsg}{Logical. If \code{TRUE}, ...  (default: TRUE).}

\item{plot_PSdiagnostics}{Logical. If \code{TRUE}, ... (default: FALSE).}

\item{randomize}{Logical. If \code{TRUE}, randomizes treatment \code{A} (default: FALSE).}

\item{ensure_cluster_positivity}{Logical. If \code{TRUE}, will only generate data that meets the positivity assumption (default: TRUE).}
}
\value{
A list containing:
\describe{
\item{\code{data}}{A data frame containing the simulated data: \code{A} (treatment), \code{Y} (outcome),
\code{Z} (cluster-level var), \code{X1},...,\code{Xp} (individual-level covariates),
\code{school} (cluster ID), \code{W_nj} (standardized cluster size),
\code{ps_true} (true propensity score), \code{iptw_true} (IPTW weights).}
\item{\code{truevals}}{A list from \code{trueVals_noMed()} containing the true total treatment
effect at the individual and cluster levels. \code{NULL} if \code{include_truevals = FALSE}.}
\item{\code{overlap}}{A list containing diagnostic plots and summary for propensity scores (\code{ps_true})
and stabilized IPTW (\code{iptw_true}). Useful for checking overlap/outliers.}
\item{\code{parameters}}{A list of the input parameters and additional generated quantities
(e.g., \code{nj_sizes}, \code{y_given}, \code{clust_trt_prop}).}
}
}
\description{
Generates a clustered dataset with an individual-level treatment (\code{A}),
an outcome (\code{Y}), individual-level covariates (\code{X}), and a
cluster-level variable (\code{Z}). A simplified version of \code{generate_data()}
with no mediator. Supports both binary and continuous outcomes and computes
the true total treatment effect (TTE) if requested.
}
\details{
This function is a wrapper that calls:
\itemize{
\item \code{generate_clusters()}: Builds the cluster IDs and sets cluster sizes in \code{njrange}.
\item \code{generate_confounders()}: Creates individual-level confounders \code{X} (optionally correlated with \code{Z}).
\item \code{generate_treatment()}: Simulates the binary (or continuous) treatment \code{A} with optional ICC.
\item \code{generate_outcome()}: Constructs outcome \code{Y} based on \code{A}, \code{M}, \code{Z},
interactions, ICC, etc., again depending on the family chosen.
\item \code{trueVals_noMed()}: if \code{include_truevals = TRUE}.
}
Note: No mediator is generated; all mediator-related coefficients are fixed to 0 internally when calling
\code{generate_outcome()}.
}
\examples{
result <- generate_data_noMed(
  J = 50,
  njrange = c(10, 20),
  Yfamily = "gaussian"
)
str(result$data)

}
