% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mediator.R
\name{generate_mediator}
\alias{generate_mediator}
\title{generate_mediator}
\usage{
generate_mediator(
  data_list,
  nj_sizes,
  iccm = 0.2,
  num_x = 3,
  m_on_a = 0.2,
  m_on_az = 0.2,
  m_on_anj = 0.2,
  m_on_x = sqrt(0.15/num_x),
  m_on_z = sqrt(0.4),
  quadratic.M = FALSE,
  Mfamily = "binomial"
)
}
\arguments{
\item{data_list}{A list containing a data frame with variables 'school', 'A', 'Z', 'X', and 'W_nj'.}

\item{nj_sizes}{Vector of cluster sizes (length equals the number of unique schools).}

\item{iccm}{Numeric. Intra-class correlation for 'M'.}

\item{num_x}{Numeric. Number of X covariates.}

\item{m_on_a}{Numeric. Effect of 'A' on 'M'.}

\item{m_on_az}{Numeric. Interaction effect of 'A' and 'Z' on 'M'.}

\item{m_on_anj}{Numeric. Interaction effect of 'A' and cluster size on 'M'.}

\item{m_on_x}{description}

\item{m_on_z}{description}

\item{quadratic.M}{Logical. If TRUE, includes quadratic terms for 'X' in the model for 'M'.}

\item{Mfamily}{Character. Either "binomial" or "gaussian". Determines the distribution of the mediator 'M'.}
}
\value{
The modified \code{data_list} with the generated mediator 'M' and associated parameters.
}
\description{
Generates an individual-level mediator variable (\code{M}) in a clustered setting
that is either binomial (\code{Mfamily} = "binomial") or gaussian (\code{Mfamily} = "gaussian"),
influenced by treatment, covariates, cluster-level confounder (\code{Z}), and
interactions between treatment and cluster size as well as between treatment
and cluster-level confounder.
}
\details{
Additional details about the function, such as implementation notes or edge cases.
}
\seealso{
\code{\link[=pm1]{pm1()}}
}
