% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_outcome.R
\name{generate_outcome}
\alias{generate_outcome}
\title{generate_outcome}
\usage{
generate_outcome(
  data_list,
  iccy = 0.2,
  yintercept = 1,
  y_on_a = 0.2,
  y_on_m = 1,
  y_on_am = 0,
  y_on_az = 0.2,
  y_on_mz = 0.2,
  y_on_anj = 0.2,
  num_x = 3,
  y_on_x = sqrt(0.15/num_x),
  y_on_z = sqrt(0.4),
  quadratic.Y = FALSE,
  Yfamily = "gaussian",
  if.null = FALSE
)
}
\arguments{
\item{data_list}{A list containing at least the element \code{data}, which is
a \code{data.frame} with variables \code{X}, \code{Z}, and \code{school}......NEED MORE DESCRIPTION}

\item{iccy}{Intra-class correlation for 'Y'.... description}

\item{yintercept}{Numeric. Intercept in the outcome \code{Y} model (default: 1).}

\item{y_on_a}{Numeric. Main effect of \code{A} on \code{Y} (default: 0.5).}

\item{y_on_m}{Numeric. Main effect of \code{M} on \code{Y} (default: 1).}

\item{y_on_am}{Numeric. Interaction effect of \code{A} and \code{M} on \code{Y} (default: 0).}

\item{y_on_az}{Numeric. Interaction effect of \code{A} and \code{Z} on \code{Y} (default: 0.2).}

\item{y_on_mz}{Numeric. Interaction effect of \code{M} and \code{Z} on \code{Y} (default: 0.2).}

\item{y_on_anj}{Numeric. Interaction effect of \code{A} and cluster size (\code{nj}) on \code{Y} (default: 0.2).}

\item{num_x}{Numeric. Number of X covariates.}

\item{y_on_x}{Numeric. Effect of each individual-level covariate \code{X} on \code{Y}.
Defaults to \code{sqrt(0.15 / num_x)}.}

\item{y_on_z}{Numeric. Effect of \code{Z} on \code{Y}. Defaults to \code{sqrt(0.4)}.}

\item{quadratic.Y}{Logical. If \code{TRUE}, include quadratic term for \code{Y} itself (used in some advanced simulations) (default: FALSE).}

\item{Yfamily}{Character. Family for the outcome (\code{"gaussian"} or \code{"binomial"}). Defaults to \code{"binomial"}.}

\item{if.null}{Logical. If \code{TRUE}, generate data under the null hypothesis (i.e., no effects). Defaults to \code{FALSE}.}
}
\value{
Description of the output value or object the function returns.
}
\description{
Brief description of what the function does.
}
\details{
Additional details about the function, such as implementation notes or edge cases.
}
\seealso{
\code{\link[=my]{my()}}
}
