% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_treatment.R
\name{generate_treatment}
\alias{generate_treatment}
\title{generate_treatment}
\usage{
generate_treatment(
  data_list,
  nj_sizes,
  icca,
  quadratic.A = FALSE,
  a_on_x = sqrt(0.05625/3),
  a_on_z = sqrt(0.15/1),
  num_x = 3,
  randomize = FALSE
)
}
\arguments{
\item{data_list}{A list containing at least the element \code{data}, which is
a \code{data.frame} with variables \code{X}, \code{Z}, and \code{school}.}

\item{nj_sizes}{A numeric vector of length \code{J} (number of schools)
indicating the number of individuals in each school.}

\item{icca}{A numeric value for the intra-class correlation for \code{A}.}

\item{quadratic.A}{Logical. If \code{TRUE}, quadratic terms for \code{X}
(transformed as (\code{X}^2 - 1)/sqrt(4)) are used in the treatment model.
Default is \code{FALSE}.}

\item{a_on_x}{Numeric; coefficient on \code{rowSums(X)} in the treatment model.}

\item{a_on_z}{Numeric; coefficient on \code{Z} in the treatment model.}

\item{num_x}{Number of columns in \code{X}. Default is 3.}

\item{randomize}{Logical. If \code{TRUE}, assignments are completely
randomized with probability 0.5 for treatment vs control, overriding the
model-based propensity score. Default is \code{FALSE}.}
}
\value{
Returns \code{data_list} with updated elements:
\itemize{
\item \code{data$A}: treatment assignments
\item \code{data$ps_true}: true propensity scores
\item \code{a_on_x}, \code{a_on_z}: the coefficients used to generate
\item \code{icca}, \code{quadratic.A}: the intra-correlation coefficient for treatment and whether pretreatment covariates are quadraticly related to A
\item \code{clust_trt_prop}: cluster-level treatment proportions
}
}
\description{
Generates a binary, individual-level treatment (\code{A}) in a clustered setting
either by using a propensity score (default; influenced by \code{Z} and \code{X})
or by completely random assignment (when randomize = \code{TRUE}; i.e., randomized experiment).
}
