% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.R
\name{my}
\alias{my}
\title{\code{my()} generates clustered data consisting of a level-1 treatment,
3 level-1 confounders, a level-1 mediator, and a level-1 outcome as well as
a level-2 confounder, with control over the number of clusters, cluster size,
and ICC. Specifically, the dataframe returned from this function consists of
the following variables: observation ID (id); cluster ID (school);
3 level-1 confounders of T, M, Y relations (x1-x3); a level-2 confounder of
T, M, Y relations (z); (t_ast); the true propensity score of an observation (ps_true);
a level-1 treatment assignment (t); level-1 mediator value (m); and the level-1 outcome measure (y).}
\usage{
my(m, a, z, nj, given, gen_y, binary = TRUE)
}
\arguments{
\item{m}{description........NEED UPDATE}

\item{a}{description........NEED UPDATE}

\item{z}{description........NEED UPDATE}

\item{nj}{description........NEED UPDATE}

\item{given}{description........NEED UPDATE}

\item{gen_y}{description........NEED UPDATE}

\item{binary}{Logical. If TRUE, ....description........NEED UPDATE}
}
\value{
Returns a dataframe of generated data
}
\description{
\code{my()} generates clustered data consisting of a level-1 treatment,
3 level-1 confounders, a level-1 mediator, and a level-1 outcome as well as
a level-2 confounder, with control over the number of clusters, cluster size,
and ICC. Specifically, the dataframe returned from this function consists of
the following variables: observation ID (id); cluster ID (school);
3 level-1 confounders of T, M, Y relations (x1-x3); a level-2 confounder of
T, M, Y relations (z); (t_ast); the true propensity score of an observation (ps_true);
a level-1 treatment assignment (t); level-1 mediator value (m); and the level-1 outcome measure (y).
}
