% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trueVals2.0f.R
\name{trueVals2.0f}
\alias{trueVals2.0f}
\title{trueVals2.0f}
\usage{
trueVals2.0f(data_list, from_GenData_2.0 = FALSE)
}
\arguments{
\item{data_list}{A list containing the generated dataset (\code{data}), cluster sizes,
and all generation parameters. This should be the output from one
of the data generation functions (e.g., \code{generate_data()}).}

\item{from_GenData_2.0}{Logical. if FALSE, ...... description}
}
\value{
A list with two elements:
\describe{
\item{\code{truevals_individual}}{A named list of true expected outcomes at the individual level
for each scenario \code{Y(a0, gm(a1))}.}
\item{\code{truevals_cluster}}{A named list of true expected outcomes at the cluster level for
each scenario \code{Y(a0, gm(a1))}.}
}
}
\description{
Computes the true potential outcomes under interventions on treatment (\code{A})
and mediator (\code{M}) for the given data configuration. It returns these true
values at both the individual and cluster levels, enabling the calculation of
natural direct/indirect effects in a clustered setting.
}
\details{
This function calculates the expected outcomes under the four potential-outcome scenarios:
\itemize{
\item \eqn{Y(a_0 = 0, gm(a_1 = 0))}: Set \code{A = 0} and \code{M} to the distribution
it would have under \code{A = 0}.
\item \eqn{Y(a_0 = 0, gm(a_1 = 1))}: Set \code{A = 0} and \code{M} to the distribution
it would have under \code{A = 1}.
\item \eqn{Y(a_0 = 1, gm(a_1 = 0))}: Set \code{A = 1} and \code{M} to the distribution
it would have under \code{A = 0}.
\item \eqn{Y(a_0 = 1, gm(a_1 = 1))}: Set \code{A = 1} and \code{M} to the distribution
it would have under \code{A = 1}.
}

For \strong{binary mediators}, the function sums over \code{M = 0, 1} with respective probabilities.
For \strong{continuous mediators}, it integrates (via closed-form or Monte Carlo) over the distribution of \code{M}.
For \strong{binary outcomes}, \code{Y} is derived from a probit link, i.e. \code{pnorm(latentY)}.
For \strong{continuous outcomes}, the linear predictor \code{latentY} is used directly.

Additional logic is included for the special case of a \strong{continuous mediator} combined with a
\strong{binary outcome}, where a large population dataset is generated (by calling \code{generate_data()}
internally) to approximate the integral. This provides a Monte Carlo estimate of the expected
outcome \eqn{E[Y]}.
}
\examples{
\dontrun{
# Typically called internally. Example call:
data_list <- generate_data(J = 50, Mfamily = "binomial", Yfamily = "binomial")
result <- trueVals2.0f(data_list)
result$truevals_individual
}

}
\seealso{
\code{\link{generate_data}}
}
